refactor(web): colocate not-interested filter predicate in useCollections
